requirements.txt
django
djangorestframework
django-simple-history